requirements.txt
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
greenlet==2.0.1
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==5.1.5
kiwisolver==1.4.4
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
six==1.16.0
SQLAlchemy==1.4.46
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
wcwidth==0.2.6
Yaml8==0.1.2
